Reorg asset paths service
