+++ dandiapi/api/services/asset/paths/query.py
