+++ dandiapi/api/services/asset/paths/utils.py
